+++ Changelog.md
@@ -6,6 +6,7 @@
 - Only dispatch to the background queue if needed
 - Correctly handle cancellation of delayed responses
 - Adding and reading mocks is now thread safe by using a Dispatch Semaphore
+- Add support for using Swift Package Manager
 
 ### 1.3.0
 - Updated to Swift 4.2